Move utils.code_analysis module to the Editor plugin
