+++ spyder/plugins/editor/utils/code_analysis/__init__.py
