+++ spyder/plugins/editor/utils/code_analysis/decorators.py
