+++ spyder/plugins/editor/utils/code_analysis/lsp_transport/consumer.py
